projects/01-first-project/projects/03-evaluated-rag/projects/03-evaluated-rag/requirements.txt
sentence-transformers==3.1.1
faiss-cpu
pypdf==5.0.1
gradio==5.9.1
ollama==0.4.3
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
numpy
pandas